+++ NEWS.md
@@ -1,5 +1,10 @@
 # SF6Dvalues 0.2.0.9000
 
+* New `SF36` class for SF-36 responses.
+
+* New `SF12_scores()` calculates SF-12 health domain scale scores (50/10
+  norm-based).
+
 # SF6Dvalues 0.2.0
 
 * New `PCS()` and `MCS()` calculate SF-12 Component Summary scores.
@@ -7,6 +12,6 @@
 # SF6Dvalues 0.1.0
 
 * First release: contains `SF12` and `SF6D` classes and `sf6d_utility()` and
-    `utility()` to calculate utility values.
+  `utility()` to calculate utility values.
 
 * Added a `NEWS.md` file to track changes to the package.